// frontend/src/common/enums/language.enum.ts
export enum LanguageProficiency {
  BASIC = 'BASIC',
  INTERMEDIATE = 'INTERMEDIATE',
  FLUENT = 'FLUENT',
}

export enum LanguageProficiencyName {
  BASIC = 'Basic',
  INTERMEDIATE = 'Intermediate',
  FLUENT = 'Fluent',
}

export enum LanguageLevelType {
  VERBAL = 'VERBAL',
  WRITTEN = 'WRITTEN',
  BOTH = 'BOTH',
}

export enum LanguageLevelTypeName {
  VERBAL = 'Verbal',
  WRITTEN = 'Written',
  BOTH = 'Both',
}